+++ sdk/nodejs/state/index.ts
@@ -26,3 +26,4 @@ export * from "./remoteBackendConfig"
 export * from "./remoteStateReference";
 export * from "./s3BackendConfig";
 export * from "./swiftBackendConfig"
+export * from "./ossBackendConfig"